requirements.txt
Django==2.1.3
djangorestframework==3.9.0
pytz==2018.7
dj-database-url==0.5.0
djangorestframework-gis==0.13
psycopg2<2.7.4
